+++ core/github.py
@@ -23,25 +23,39 @@ SOFTWARE.
 """
 
 
-from urllib.request import urlopen
+import aiohttp
 
 
-def get_latest():
+async def get_latest():
     # Get latest version string from GitHub
-    data = urlopen(
-        "https://raw.githubusercontent.com/eibex/reaction-light/master/.version"
-    )
-    for line in data:
-        latest = line.decode()
-        break
+    async with aiohttp.ClientSession() as cs:
+        async with cs.get(
+            "https://raw.githubusercontent.com/eibex/reaction-light/master/.version"
+        ) as r:
+            latest = await r.text()
 
     return latest.rstrip("\n").rstrip("\r")
 
 
-def check_for_updates(version):
+async def check_for_updates(version):
     # Get latest version from GitHub repo and checks it against the current one
-    latest = get_latest()
+    latest = await get_latest()
     if latest > version:
         return latest
 
     return False
+
+
+async def latest_changelog():
+    # Get the changes for the latest version
+    async with aiohttp.ClientSession() as cs:
+        async with cs.get(
+            "https://raw.githubusercontent.com/eibex/reaction-light/master/CHANGELOG.md"
+        ) as r:
+            changelog = await r.text()
+
+    changelog = changelog.split("###")[1].rstrip("\n")  # Only get the latest version changes
+    changelog = changelog[changelog.index("-"):]  # Remove every character up to the first bullet point
+    changelog = changelog + "\n\n[View more](https://github.com/eibex/reaction-light/blob/master/CHANGELOG.md)"
+
+    return changelog